WalltechOLEDWatchv4.3
This program is 1k oversized and needs compressing somehow. This will be the non-interrupt version that will have a 2 day battery life like previous versions. When the "temperature" face is selected, the icons shows, the 1 second delay passes, and the sketch either resets, or hangs completely. I assume it's a memory issue from experience and forum browsing.